Move statuses.py and workload.py from common to core.
