+++ charmcraft.yaml
@@ -6,7 +6,6 @@ platforms:
   ubuntu@24.04:amd64:
 
 name: cassandra
-display-name: Cassandra VM
 summary: |
   Machine charm for Cassandra
 description: |
